requirements.txt
Django==2.0.2
Pillow==5.0.0
pytz==2018.3
